+++ et/audio/modules/module.cc
@@ -3,8 +3,8 @@
 namespace Et {
 namespace Audio {
 
-Module::Module(unsigned int sampleRate, unsigned int bufferSize, unsigned int nInputs,
-               unsigned int nParameters)
+Module::Module(unsigned int sampleRate, size_t bufferSize, size_t nInputs,
+               size_t nParameters)
     : on_{true}
     , sampleRate_{sampleRate}
     , bufferSize_{bufferSize}
@@ -12,12 +12,12 @@ Module::Module(unsigned int sampleRate, unsigned int bufferSize, unsigned int nI
     , output_(*this)
 {
     inputs_.reserve(nInputs);
-    for(unsigned int i=0; i<nInputs; ++i) {
+    for(size_t i=0; i<nInputs; ++i) {
         inputs_.push_back(std::move(Input(*this)));
     }
     
     params_.reserve(nParameters);
-    for(unsigned int i=0; i<nParameters; ++i) {
+    for(size_t i=0; i<nParameters; ++i) {
         params_.push_back(std::move(Parameter(*this, -1.0f, 1.0f, 0.0f)));
     }
 }
